Restructure/Organize pkg/dashboard into sub-packages
