+++ pkg/dashboard/dashboard.go
@@ -15,62 +15,18 @@
 package dashboard
 
 import (
-	"bytes"
-	"encoding/json"
-	"html/template"
+	"fmt"
 	"net/http"
-	"strings"
 
 	packr "github.com/gobuffalo/packr/v2"
 	"github.com/gorilla/mux"
-	"k8s.io/apimachinery/pkg/util/sets"
-	"k8s.io/klog"
 
-	"github.com/fairwindsops/goldilocks/pkg/summary"
-)
-
-const (
-	// MainTemplateName is the main template
-	MainTemplateName = "main.gohtml"
-	// HeadTemplateName contains styles and meta info
-	HeadTemplateName = "head.gohtml"
-	// NavbarTemplateName contains the navbar
-	NavbarTemplateName = "navbar.gohtml"
-	// PreambleTemplateName contains an empty preamble that can be overridden
-	PreambleTemplateName = "preamble.gohtml"
-	// DashboardTemplateName contains the content of the dashboard
-	DashboardTemplateName = "dashboard.gohtml"
-	// NamespaceTemplateName contains the content for a namespace
-	NamespaceTemplateName = "namespace.gohtml"
-	// ContainerTemplateName contains the content for a container
-	ContainerTemplateName = "container.gohtml"
-	// FooterTemplateName contains the footer
-	FooterTemplateName = "footer.gohtml"
-	// CheckDetailsTemplateName is a page for rendering details about a given check
-	CheckDetailsTemplateName = "check-details.gohtml"
+	"github.com/fairwindsops/goldilocks/pkg/dashboard/handlers"
 )
 
 var (
-	templateBox = (*packr.Box)(nil)
-	assetBox    = (*packr.Box)(nil)
 	markdownBox = (*packr.Box)(nil)
 )
-
-// GetAssetBox returns a binary-friendly set of assets packaged from disk
-func GetAssetBox() *packr.Box {
-	if assetBox == (*packr.Box)(nil) {
-		assetBox = packr.New("Assets", "assets")
-	}
-	return assetBox
-}
-
-// GetTemplateBox returns a binary-friendly set of templates for rendering the dash
-func GetTemplateBox() *packr.Box {
-	if templateBox == (*packr.Box)(nil) {
-		templateBox = packr.New("Templates", "templates")
-	}
-	return templateBox
-}
 
 // GetMarkdownBox returns a binary-friendly set of markdown files with error details
 func GetMarkdownBox() *packr.Box {
@@ -80,130 +36,32 @@ func GetMarkdownBox() *packr.Box {
 	return markdownBox
 }
 
-// templateData is passed to the dashboard HTML template
-type templateData struct {
-	BasePath string
-	VPAData  summary.Summary
-	JSON     template.JS
-}
-
-// GetBaseTemplate puts together the dashboard template. Individual pieces can be overridden before rendering.
-func GetBaseTemplate(name string) (*template.Template, error) {
-	tmpl := template.New(name).Funcs(template.FuncMap{
-		"printResource":  printResource,
-		"getStatus":      getStatus,
-		"getStatusRange": getStatusRange,
-		"resourceName":   resourceName,
-		"getUUID":        getUUID,
-	})
-
-	templateFileNames := []string{
-		DashboardTemplateName,
-		NamespaceTemplateName,
-		ContainerTemplateName,
-		HeadTemplateName,
-		NavbarTemplateName,
-		PreambleTemplateName,
-		FooterTemplateName,
-		MainTemplateName,
-	}
-	return parseTemplateFiles(tmpl, templateFileNames)
-}
-
-func parseTemplateFiles(tmpl *template.Template, templateFileNames []string) (*template.Template, error) {
-	templateBox := GetTemplateBox()
-	for _, fname := range templateFileNames {
-		templateFile, err := templateBox.Find(fname)
-		if err != nil {
-			return nil, err
-		}
-
-		tmpl, err = tmpl.Parse(string(templateFile))
-		if err != nil {
-			return nil, err
-		}
-	}
-	return tmpl, nil
-}
-
-func writeTemplate(tmpl *template.Template, data *templateData, w http.ResponseWriter) {
-	buf := &bytes.Buffer{}
-	err := tmpl.Execute(buf, data)
-	if err != nil {
-		klog.Errorf("Error executing template: %v", err)
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	_, err = buf.WriteTo(w)
-	if err != nil {
-		klog.Errorf("Error writing template: %v", err)
-	}
-}
-
 // GetRouter returns a mux router serving all routes necessary for the dashboard
 func GetRouter(port int, basePath string, vpaLabels map[string]string, excludeContainers string) *mux.Router {
 	router := mux.NewRouter()
-	router.HandleFunc("/health", func(w http.ResponseWriter, r *http.Request) {
-		_, err := w.Write([]byte("OK"))
-		if err != nil {
-			klog.Errorf("Error writing healthcheck: %v", err)
-		}
-	})
-	router.HandleFunc("/favicon.ico", func(w http.ResponseWriter, r *http.Request) {
-		favicon, err := GetAssetBox().Find("images/favicon-32x32.png")
-		if err != nil {
-			klog.Errorf("Error getting favicon: %v", err)
-			http.Error(w, "Error getting favicon", http.StatusInternalServerError)
-			return
-		}
-		_, err = w.Write(favicon)
-		if err != nil {
-			klog.Errorf("Error writing favicon: %v", err)
-		}
-	})
 
-	fileServer := http.FileServer(GetAssetBox())
-	router.PathPrefix("/static/").Handler(http.StripPrefix("/static/", fileServer))
+	// health
+	router.Handle("/health", handlers.Health("OK"))
+	router.Handle("/healthz", handlers.Healthz())
+
+	// assets
+	router.Handle("/favicon.ico", handlers.Asset("images/favicon-32x32.png"))
+	router.PathPrefix("/static/").Handler(handlers.StaticAssets("/static/"))
+
+	// dashboard
+	router.Handle("/dashboard", handlers.DashboardAllNamespaces(basePath, vpaLabels, excludeContainers))
+
+	// root
 	router.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
+		// catch all other paths that weren't matched
 		if r.URL.Path != "/" {
 			http.NotFound(w, r)
 			return
 		}
 
-		summarizer := summary.NewSummarizer(
-			summary.ForVPAsWithLabels(vpaLabels),
-			summary.ExcludeContainers(sets.NewString(strings.Split(excludeContainers, ",")...)),
-		)
-
-		data, err := summarizer.GetSummary()
-		if err != nil {
-			klog.Errorf("Error getting data: %v", err)
-			http.Error(w, "Error running summary.", 500)
-			return
-		}
-		MainHandler(w, r, data, basePath)
+		// default redirect on root path
+		http.Redirect(w, r, fmt.Sprintf("%s://%s/dashboard", r.URL.Scheme, r.URL.Host), http.StatusMovedPermanently)
 	})
+
 	return router
-}
-
-// MainHandler gets template data and renders the dashboard with it.
-func MainHandler(w http.ResponseWriter, r *http.Request, vpaData summary.Summary, basePath string) {
-	jsonData, err := json.Marshal(vpaData)
-	if err != nil {
-		http.Error(w, "Error serializing summary data", 500)
-		return
-	}
-
-	data := templateData{
-		BasePath: basePath,
-		VPAData:  vpaData,
-		JSON:     template.JS(jsonData),
-	}
-	tmpl, err := GetBaseTemplate("main")
-	if err != nil {
-		klog.Errorf("Error getting template data %v", err)
-		http.Error(w, "Error getting template data", 500)
-		return
-	}
-	writeTemplate(tmpl, &data, w)
 }
